Drop double useQuoteDeposit subscription
Move ownership of the `useQuoteDeposit` / `useQuoteRedeem`
queries from the fees hooks to the forms. `useDepositFees` and
`useWithdrawFees` now take `quote` + `shares` as inputs, derive
`canSubmit` and the slippage minimum internally, and return them.

The deposit form previously called `useQuoteDeposit` standalone
AND `useDepositFees` called it again internally; the duplicate
subscription is gone now that the form is the single owner. Both
hooks share a mirror shape, keeping the deposit and withdraw
forms structurally symmetric.
